@@ -1,6 +1,6 @@
 {
   "name": "@devias/carpatin-dashboard-free",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "license": "MIT",
   "scripts": {
     "start": "react-scripts start",
@@ -20,34 +20,28 @@
       "last 1 safari version"
     ]
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
   "dependencies": {
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/material": "^5.8.1",
-    "@mui/icons-material": "^5.8.0",
-    "@mui/lab": "^5.0.0-alpha.83",
-    "apexcharts": "^3.35.3",
-    "date-fns": "^2.28.0",
-    "formik": "^2.2.9",
-    "history": "^5.3.0",
-    "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-apexcharts": "^1.4.0",
-    "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "^5.0.1",
-    "simplebar": "^5.3.8",
-    "simplebar-react": "^2.4.1",
-    "yup": "^0.32.11"
+    "@emotion/react": "11.10.6",
+    "@emotion/styled": "11.10.6",
+    "@heroicons/react": "2.0.16",
+    "@mui/icons-material": "5.11.9",
+    "@mui/lab": "5.0.0-alpha.120",
+    "@mui/material": "5.11.10",
+    "apexcharts": "3.37.1",
+    "date-fns": "2.29.3",
+    "formik": "2.2.9",
+    "numeral": "2.0.6",
+    "prop-types": "15.8.1",
+    "react": "18.2.0",
+    "react-apexcharts": "1.4.0",
+    "react-dom": "18.2.0",
+    "react-helmet-async": "1.3.0",
+    "react-router-dom": "6.8.1",
+    "simplebar-react": "3.2.1",
+    "yup": "1.0.0"
   },
   "devDependencies": {
-    "eslint-config-airbnb": "^18.2.1"
+    "eslint": "8.34.0",
+    "react-scripts": "5.0.1"
   }
 }
