@@ -1,17 +1,20 @@
 import { useRoutes } from 'react-router-dom';
 import { CssBaseline, ThemeProvider } from '@mui/material';
-import { theme } from './theme';
 import { routes } from './routes';
+import { createTheme } from './theme';
+import 'simplebar-react/dist/simplebar.min.css';
 
-const App = () => {
-  const content = useRoutes(routes);
+export const App = () => {
+  const element = useRoutes(routes);
+  const theme = createTheme({
+    colorPreset: 'green',
+    contrast: 'high'
+  });
 
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
-      {content}
+      {element}
     </ThemeProvider>
   );
 };
-
-export default App;
