@@ -1,66 +1,40 @@
-export const Logo = () => (
-  <svg
-    width="117"
-    height="24"
-    viewBox="0 0 117 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <g clipPath="url(#clip0)">
+import PropTypes from 'prop-types';
+import { useTheme } from '@mui/material/styles';
+
+export const Logo = (props) => {
+  const { color: colorProp = 'primary' } = props;
+  const theme = useTheme();
+
+  const color = colorProp === 'primary'
+    ? theme.palette.primary.main
+    : colorProp === 'black'
+      ? '#1D262D' : '#FFFFFF';
+
+  return (
+    <svg
+      fill="none"
+      height="100%"
+      viewBox="0 0 24 25"
+      width="100%"
+      xmlns="http://www.w3.org/2000/svg"
+    >
       <path
-        fillRule="evenodd"
         clipRule="evenodd"
-        d="M20.8963 14.1031C22.7584 14.1031 23.97 16.0899 23.0899 17.7308C21.019 21.5913 16.8744 24.1739 12.1469 23.9908C5.993 23.7526 0.981706 18.7348 0.750861 12.5807C0.497005 5.81318 5.90789 0.247406 12.6187 0.247406C17.132 0.247406 21.0588 2.76642 23.0674 6.47504C23.9641 8.1308 22.7796 10.1443 20.8968 10.1443C19.9998 10.1443 19.1579 9.66753 18.7363 8.87578C17.4672 6.4931 14.8441 4.93459 11.9012 5.23224C8.57114 5.56923 5.92892 8.3119 5.70649 11.6516C5.43779 15.6846 8.64215 19.0515 12.6187 19.0515C15.2538 19.0515 17.5514 17.5717 18.7207 15.4003C19.1517 14.6004 19.9877 14.1031 20.8963 14.1031Z"
-        fill="white"
+        d="M20.8962 14.6031C22.7584 14.6031 23.97 16.5899 23.0899 18.2308C21.019 22.0913 16.8744 24.6739 12.1469 24.4908C5.99299 24.2526 0.981695 19.2348 0.750849 13.0807C0.496994 6.3132 5.90787 0.747423 12.6187 0.747423C17.132 0.747423 21.0588 3.26644 23.0674 6.97506C23.964 8.63081 22.7796 10.6443 20.8967 10.6443C19.9998 10.6443 19.1579 10.1676 18.7362 9.3758C17.4672 6.99312 14.844 5.4346 11.9012 5.73225C8.57113 6.06924 5.92891 8.81192 5.70648 12.1516C5.43778 16.1846 8.64214 19.5515 12.6187 19.5515C15.2538 19.5515 17.5513 18.0717 18.7207 15.9003C19.1517 15.1004 19.9877 14.6031 20.8962 14.6031Z"
+        fill={color}
+        fillRule="evenodd"
       />
       <path
+        clipRule="evenodd"
+        d="M16.5774 12.6237C16.5774 14.8102 14.8049 16.5825 12.6187 16.5825C10.4322 16.5825 8.65991 14.8102 8.65991 12.6237C8.65991 10.4372 10.4322 8.66495 12.6187 8.66495C14.8049 8.66495 16.5774 10.4372 16.5774 12.6237Z"
+        fill={color}
+        fillRule="evenodd"
         opacity="0.5"
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M16.5774 12.1237C16.5774 14.3102 14.8049 16.0825 12.6187 16.0825C10.4322 16.0825 8.65991 14.3102 8.65991 12.1237C8.65991 9.93724 10.4322 8.16495 12.6187 8.16495C14.8049 8.16495 16.5774 9.93724 16.5774 12.1237Z"
-        fill="white"
       />
-      <path
-        d="M37.1279 4.14488C36.9489 2.85226 35.9546 2.10226 34.6648 2.10226C33.1449 2.10226 32 3.21022 32 5.0909C32 6.96874 33.1307 8.07954 34.6648 8.07954C36.0483 8.07954 36.966 7.18465 37.1279 6.07385L36.0654 6.06817C35.9319 6.74999 35.3751 7.13351 34.679 7.13351C33.7358 7.13351 33.0597 6.42613 33.0597 5.0909C33.0597 3.7784 33.7302 3.04829 34.6819 3.04829C35.3893 3.04829 35.9432 3.44885 36.0654 4.14488H37.1279ZM40.7699 7.99999L41.2501 6.56533H43.4375L43.9205 7.99999H45.0455L42.9944 2.18181H41.6932L39.6449 7.99999H40.7699ZM41.5341 5.71874L42.3211 3.37499H42.3665L43.1534 5.71874H41.5341ZM47.8118 7.99999H48.8658V5.86931H49.9397L51.0817 7.99999H52.2578L50.9993 5.69317C51.684 5.4176 52.0448 4.83238 52.0448 4.0426C52.0448 2.93181 51.3289 2.18181 49.9936 2.18181H47.8118V7.99999ZM48.8658 4.99431V3.06249H49.8317C50.6186 3.06249 50.9709 3.42329 50.9709 4.0426C50.9709 4.66192 50.6186 4.99431 49.8374 4.99431H48.8658ZM55.015 7.99999H56.0689V6.03408H57.1826C58.5263 6.03408 59.2479 5.22726 59.2479 4.10794C59.2479 2.99715 58.5348 2.18181 57.1968 2.18181H55.015V7.99999ZM56.0689 5.1676V3.06249H57.0348C57.8246 3.06249 58.174 3.48863 58.174 4.10794C58.174 4.72726 57.8246 5.1676 57.0405 5.1676H56.0689ZM62.2621 7.99999L62.7422 6.56533H64.9297L65.4127 7.99999H66.5377L64.4865 2.18181H63.1854L61.1371 7.99999H62.2621ZM63.0263 5.71874L63.8132 3.37499H63.8587L64.6456 5.71874H63.0263ZM68.3452 3.06533H70.1435V7.99999H71.189V3.06533H72.9873V2.18181H68.3452V3.06533ZM76.9361 2.18181H75.8821V7.99999H76.9361V2.18181ZM84.8587 2.18181H83.8104V6.14772H83.7593L81.0206 2.18181H80.0775V7.99999H81.1314V4.03692H81.1797L83.9212 7.99999H84.8587V2.18181Z"
-        fill="white"
-      />
-      <path
-        d="M35.2961 23.5895C38.3984 23.5895 40.1931 21.6655 40.1931 18.4837C40.1931 15.3168 38.3984 13.4077 35.4005 13.4077H32V23.5895H35.2961ZM33.5362 22.2472V14.75H35.3061C37.5333 14.75 38.6818 16.0824 38.6818 18.4837C38.6818 20.8949 37.5333 22.2472 35.2116 22.2472H33.5362ZM44.3565 23.7585C45.6192 23.7585 46.3302 23.1172 46.6136 22.5455H46.6732V23.5895H48.1249V18.5185C48.1249 16.2962 46.3749 15.8537 45.1619 15.8537C43.7798 15.8537 42.507 16.4105 42.0099 17.8026L43.4069 18.1207C43.6256 17.5788 44.1825 17.0568 45.1817 17.0568C46.1413 17.0568 46.6334 17.5589 46.6334 18.424V18.4588C46.6334 19.0007 46.0766 18.9908 44.7045 19.1499C43.2577 19.3189 41.7762 19.6967 41.7762 21.4318C41.7762 22.9333 42.9048 23.7585 44.3565 23.7585ZM44.6796 22.5653C43.8394 22.5653 43.2329 22.1875 43.2329 21.4517C43.2329 20.6562 43.9388 20.3729 44.7989 20.2585C45.2812 20.1939 46.4246 20.0646 46.6384 19.8508V20.8352C46.6384 21.7401 45.9175 22.5653 44.6796 22.5653ZM55.9772 17.8175C55.6689 16.6293 54.7392 15.8537 53.0887 15.8537C51.3635 15.8537 50.1405 16.7635 50.1405 18.1158C50.1405 19.1996 50.7968 19.9205 52.2286 20.2386L53.5212 20.522C54.257 20.6861 54.6 21.0142 54.6 21.4915C54.6 22.0831 53.9686 22.5455 52.9942 22.5455C52.1043 22.5455 51.5326 22.1626 51.3536 21.4119L49.9168 21.6307C50.1654 22.983 51.289 23.7436 53.0041 23.7436C54.8486 23.7436 56.1263 22.7642 56.1263 21.3821C56.1263 20.3033 55.4402 19.6371 54.0382 19.3139L52.8252 19.0355C51.985 18.8366 51.622 18.5533 51.627 18.0362C51.622 17.4496 52.2584 17.032 53.1036 17.032C54.0283 17.032 54.4559 17.544 54.6299 18.0561L55.9772 17.8175ZM59.397 19.0554C59.397 17.8374 60.168 17.1413 61.227 17.1413C62.251 17.1413 62.862 17.7926 62.862 18.9112V23.5895H64.349V18.7322C64.349 16.8281 63.305 15.8537 61.734 15.8537C60.546 15.8537 59.83 16.3707 59.472 17.196H59.377V13.4077H57.9107V23.5895H59.397V19.0554ZM66.602 23.5895H68.054V22.4013H68.178C68.446 22.8885 68.993 23.7386 70.385 23.7386C72.235 23.7386 73.577 22.2571 73.577 19.7862C73.577 17.3104 72.215 15.8537 70.37 15.8537C68.953 15.8537 68.441 16.7187 68.178 17.191H68.088V13.4077H66.602V23.5895ZM68.059 19.7713C68.059 18.1754 68.755 17.1165 70.052 17.1165C71.399 17.1165 72.076 18.255 72.076 19.7713C72.076 21.3026 71.38 22.4709 70.052 22.4709C68.774 22.4709 68.059 21.3771 68.059 19.7713ZM78.613 23.7436C80.766 23.7436 82.172 22.1676 82.172 19.8061C82.172 17.4297 80.766 15.8537 78.613 15.8537C76.46 15.8537 75.053 17.4297 75.053 19.8061C75.053 22.1676 76.46 23.7436 78.613 23.7436ZM78.618 22.4957C77.211 22.4957 76.555 21.2678 76.555 19.8011C76.555 18.3395 77.211 17.0966 78.618 17.0966C80.015 17.0966 80.671 18.3395 80.671 19.8011C80.671 21.2678 80.015 22.4957 78.618 22.4957ZM86.204 23.7585C87.467 23.7585 88.178 23.1172 88.461 22.5455H88.521V23.5895H89.972V18.5185C89.972 16.2962 88.223 15.8537 87.009 15.8537C85.627 15.8537 84.355 16.4105 83.857 17.8026L85.254 18.1207C85.473 17.5788 86.03 17.0568 87.029 17.0568C87.989 17.0568 88.481 17.5589 88.481 18.424V18.4588C88.481 19.0007 87.924 18.9908 86.552 19.1499C85.105 19.3189 83.624 19.6967 83.624 21.4318C83.624 22.9333 84.752 23.7585 86.204 23.7585ZM86.527 22.5653C85.687 22.5653 85.08 22.1875 85.08 21.4517C85.08 20.6562 85.786 20.3729 86.647 20.2585C87.129 20.1939 88.272 20.0646 88.486 19.8508V20.8352C88.486 21.7401 87.765 22.5653 86.527 22.5653ZM92.102 23.5895H93.589V18.9261C93.589 17.9268 94.359 17.206 95.413 17.206C95.721 17.206 96.069 17.2606 96.189 17.2955V15.8736C96.04 15.8537 95.746 15.8388 95.557 15.8388C94.662 15.8388 93.897 16.3459 93.618 17.1662H93.539V15.9531H92.102V23.5895ZM100.21 23.7386C101.602 23.7386 102.149 22.8885 102.418 22.4013H102.542V23.5895H103.994V13.4077H102.507V17.191H102.418C102.149 16.7187 101.642 15.8537 100.22 15.8537C98.376 15.8537 97.019 17.3104 97.019 19.7862C97.019 22.2571 98.356 23.7386 100.21 23.7386ZM100.539 22.4709C99.211 22.4709 98.52 21.3026 98.52 19.7713C98.52 18.255 99.196 17.1165 100.539 17.1165C101.836 17.1165 102.532 18.1754 102.532 19.7713C102.532 21.3771 101.821 22.4709 100.539 22.4709Z"
-        fill="white"
-      />
-      <rect
-        x="89.7216"
-        y="0.0909119"
-        width="27.5568"
-        height="9.81818"
-        rx="2"
-        fill="#F50057"
-      />
-      <path
-        d="M92.7216 7.90909H93.7756V5.4375H96.1932V4.55398H93.7756V2.97443H96.4489V2.09091H92.7216V7.90909Z"
-        fill="white"
-      />
-      <path
-        d="M98.4091 7.90909H99.4631V5.77841H100.537L101.679 7.90909H102.855L101.597 5.60228C102.281 5.32671 102.642 4.74148 102.642 3.95171C102.642 2.84091 101.926 2.09091 100.591 2.09091H98.4091V7.90909ZM99.4631 4.90341V2.97159H100.429C101.216 2.97159 101.568 3.33239 101.568 3.95171C101.568 4.57103 101.216 4.90341 100.435 4.90341H99.4631Z"
-        fill="white"
-      />
-      <path
-        d="M104.612 7.90909H108.419V7.02557H105.666V5.4375H108.2V4.55398H105.666V2.97443H108.396V2.09091H104.612V7.90909Z"
-        fill="white"
-      />
-      <path
-        d="M110.472 7.90909H114.278V7.02557H111.526V5.4375H114.06V4.55398H111.526V2.97443H114.256V2.09091H110.472V7.90909Z"
-        fill="white"
-      />
-    </g>
-    <defs>
-      <clipPath id="clip0">
-        <rect
-          width="117"
-          height="24"
-          fill="white"
-        />
-      </clipPath>
-    </defs>
-  </svg>
-);
+    </svg>
+  );
+};
+
+Logo.propTypes = {
+  color: PropTypes.oneOf(['black', 'primary', 'white'])
+};
