@@ -1,34 +1,4 @@
-import { forwardRef } from 'react';
-import PropTypes from 'prop-types';
 import SimpleBar from 'simplebar-react';
-import { Box } from '@mui/material';
+import { styled } from '@mui/material/styles';
 
-export const Scrollbar = forwardRef((props, ref) => {
-  const { children, ...other } = props;
-
-  const isMobile = /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(navigator.userAgent);
-
-  if (isMobile) {
-    return (
-      <Box
-        ref={ref}
-        sx={{ overflowX: 'auto' }}
-      >
-        {children}
-      </Box>
-    );
-  }
-
-  return (
-    <SimpleBar
-      ref={ref}
-      {...other}
-    >
-      {children}
-    </SimpleBar>
-  );
-});
-
-Scrollbar.propTypes = {
-  children: PropTypes.node.isRequired
-};
+export const Scrollbar = styled(SimpleBar)``;
