@@ -1,4 +1,4 @@
-import { Helmet } from 'react-helmet';
+import { Helmet } from 'react-helmet-async';
 import { useFormik } from 'formik';
 import * as Yup from 'yup';
 import {
@@ -8,257 +8,223 @@ import {
   Card,
   Container,
   FormHelperText,
-  Grid,
   MenuItem,
+  Stack,
   TextField,
-  Typography
+  Typography,
+  Unstable_Grid2 as Grid
 } from '@mui/material';
 
 const companySizeOptions = ['1-10', '11-30', '31-50', '50+'];
 
-export const Settings = () => {
+const initialValues = {
+  companyName: 'ACME Corp LLC.',
+  companySize: '1-10',
+  email: 'chen.simmons@acmecorp.com',
+  name: 'Chen Simmons',
+  jobTitle: 'Operation',
+  submit: null
+};
+
+const validationSchema = Yup.object({
+  companyName: Yup
+    .string()
+    .max(255)
+    .required('Company name is required'),
+  companySize: Yup
+    .string()
+    .max(255)
+    .oneOf(companySizeOptions)
+    .required('Company size is required'),
+  email: Yup
+    .string()
+    .email('Must be a valid email')
+    .max(255)
+    .required('Email is required'),
+  name: Yup
+    .string()
+    .max(255)
+    .required('Name is required'),
+  jobTitle: Yup
+    .string()
+    .max(255)
+    .required('Job title is required')
+});
+
+const Page = () => {
   const formik = useFormik({
-    initialValues: {
-      companyName: 'ACME Corp LLC.',
-      companySize: '1-10',
-      email: 'chen.simmons@acmecorp.com',
-      fullName: 'Chen Simmons',
-      jobTitle: 'Operation',
-      submit: null
-    },
-    validationSchema: Yup.object().shape({
-      companyName: Yup.string().max(255).required('Company name is required'),
-      companySize: Yup
-        .string()
-        .max(255)
-        .oneOf(companySizeOptions)
-        .required('Company size is required'),
-      email: Yup.string().email('Must be a valid email').max(255).required('Email is required'),
-      fullName: Yup.string().max(255).required('Full Name is required'),
-      jobTitle: Yup.string().max(255).required('Job name is required')
-    }),
+    initialValues,
+    validationSchema,
     onSubmit: async (values, helpers) => {
-      try {
-        helpers.setStatus({ success: true });
-        helpers.setSubmitting(false);
-      } catch (err) {
-        console.error(err);
-        helpers.setStatus({ success: false });
-        helpers.setErrors({ submit: err.message });
-        helpers.setSubmitting(false);
-      }
+      helpers.setStatus({ success: true });
+      helpers.setSubmitting(false);
     }
   });
 
   return (
     <>
       <Helmet>
-        <title>Settings | Carpatin Dashboard</title>
+        <title>
+          Settings | Carpatin Free
+        </title>
       </Helmet>
       <Box
         sx={{
-          backgroundColor: 'background.default',
-          pb: 3,
-          pt: 8
+          flexGrow: 1,
+          py: 8
         }}
       >
-        <Container maxWidth="lg">
-          <Typography
-            color="textPrimary"
-            sx={{ mb: 3 }}
-            variant="h4"
-          >
-            Settings
-          </Typography>
-          <Grid
-            container
-            spacing={3}
-          >
-            <Grid
-              item
-              md={4}
-              xs={12}
-            >
-              <Typography
-                color="textPrimary"
-                variant="h6"
-              >
-                Account
+        <Container maxWidth="xl">
+          <Stack spacing={3}>
+            <div>
+              <Typography variant="h4">
+                Settings
               </Typography>
-            </Grid>
-            <Grid
-              item
-              md={8}
-              xs={12}
-            >
-              <Card
-                variant="outlined"
-                sx={{ p: 3 }}
+            </div>
+            <div>
+              <Grid
+                container
+                spacing={3}
               >
-                <form onSubmit={formik.handleSubmit}>
-                  <div>
-                    <Box
-                      sx={{
-                        alignItems: 'center',
-                        display: 'flex',
-                        pb: 3
-                      }}
-                    >
-                      <Avatar
-                        src="/static/user-chen_simmons.png"
-                        sx={{
-                          height: 64,
-                          mr: 2,
-                          width: 64
-                        }}
-                      />
-                      <div>
-                        <Button
-                          color="primary"
-                          size="small"
-                          sx={{ mb: 1 }}
-                          type="button"
-                          variant="outlined"
-                        >
-                          Upload new picture
-                        </Button>
+                <Grid
+                  xs={12}
+                  md={4}
+                >
+                  <Typography variant="h6">
+                    Account
+                  </Typography>
+                </Grid>
+                <Grid
+                  xs={12}
+                  md={8}
+                >
+                  <Card sx={{ p: 3 }}>
+                    <form onSubmit={formik.handleSubmit}>
+                      <Stack
+                        alignItems="center"
+                        direction="row"
+                        spacing={2}
+                        sx={{ mb: 3 }}
+                      >
+                        <Avatar
+                          src="/assets/avatars/avatar-chen-simmons.jpg"
+                          sx={{
+                            height: 64,
+                            width: 64
+                          }}
+                        />
                         <div>
-                          <Typography
-                            color="textSecondary"
-                            variant="caption"
+                          <Button
+                            color="primary"
+                            size="small"
+                            type="button"
+                            variant="outlined"
                           >
-                            Recommended dimensions: 200x200, maximum file size: 5MB
-                          </Typography>
-                        </div>
-                      </div>
-                    </Box>
-                    <Grid
-                      container
-                      spacing={2}
-                      sx={{ maxWidth: 420 }}
-                    >
-                      <Grid
-                        item
-                        xs={12}
-                      >
-                        <TextField
-                          error={Boolean(formik.touched.fullName && formik.errors.fullName)}
-                          fullWidth
-                          helperText={formik.touched.fullName && formik.errors.fullName}
-                          label="Full Name"
-                          name="fullName"
-                          onBlur={formik.handleBlur}
-                          onChange={formik.handleChange}
-                          value={formik.values.fullName}
-                          variant="outlined"
-                        />
-                      </Grid>
-                      <Grid
-                        item
-                        xs={12}
-                      >
-                        <TextField
-                          error={Boolean(formik.touched.email && formik.errors.email)}
-                          fullWidth
-                          helperText={formik.touched.email && formik.errors.email}
-                          label="Email address"
-                          name="email"
-                          onBlur={formik.handleBlur}
-                          onChange={formik.handleChange}
-                          type="email"
-                          value={formik.values.email}
-                          variant="outlined"
-                        />
-                      </Grid>
-                      <Grid
-                        item
-                        xs={12}
-                      >
-                        <TextField
-                          error={Boolean(formik.touched.jobTitle && formik.errors.jobTitle)}
-                          fullWidth
-                          helperText={formik.touched.jobTitle && formik.errors.jobTitle}
-                          label="Job title"
-                          name="jobTitle"
-                          onBlur={formik.handleBlur}
-                          onChange={formik.handleChange}
-                          value={formik.values.jobTitle}
-                          variant="outlined"
-                        />
-                      </Grid>
-                      <Grid
-                        item
-                        xs={12}
-                      >
-                        <TextField
-                          error={Boolean(formik.touched.companyName && formik.errors.companyName)}
-                          fullWidth
-                          helperText={formik.touched.companyName && formik.errors.companyName}
-                          label="Company name"
-                          name="companyName"
-                          onBlur={formik.handleBlur}
-                          onChange={formik.handleChange}
-                          value={formik.values.companyName}
-                          variant="outlined"
-                        />
-                      </Grid>
-                      <Grid
-                        item
-                        xs={12}
-                      >
-                        <TextField
-                          error={Boolean(formik.touched.companySize && formik.errors.companySize)}
-                          fullWidth
-                          helperText={formik.touched.companySize && formik.errors.companySize}
-                          label="Company size"
-                          name="companySize"
-                          onBlur={formik.handleBlur}
-                          onChange={formik.handleChange}
-                          select
-                          value={formik.values.companySize}
-                          variant="outlined"
-                        >
-                          {companySizeOptions.map((companySizeOption) => (
-                            <MenuItem
-                              key={companySizeOption}
-                              value={companySizeOption}
+                            Change
+                          </Button>
+                          <div>
+                            <Typography
+                              color="text.secondary"
+                              variant="caption"
                             >
-                              {companySizeOption}
-                            </MenuItem>
-                          ))}
-                        </TextField>
-                      </Grid>
-                      {formik.errors.submit && (
-                        <Grid
-                          item
-                          xs={12}
-                        >
-                          <FormHelperText error>
+                              Recommended dimensions: 200x200, maximum file size: 5MB
+                            </Typography>
+                          </div>
+                        </div>
+                      </Stack>
+                      <Box sx={{ maxWidth: 420 }}>
+                        <Stack spacing={3}>
+                          <TextField
+                            error={Boolean(formik.touched.name && formik.errors.name)}
+                            fullWidth
+                            helperText={formik.touched.name && formik.errors.name}
+                            label="Full Name"
+                            name="name"
+                            onBlur={formik.handleBlur}
+                            onChange={formik.handleChange}
+                            value={formik.values.name}
+                          />
+                          <TextField
+                            error={Boolean(formik.touched.email && formik.errors.email)}
+                            fullWidth
+                            helperText={formik.touched.email && formik.errors.email}
+                            label="Email address"
+                            name="email"
+                            onBlur={formik.handleBlur}
+                            onChange={formik.handleChange}
+                            type="email"
+                            value={formik.values.email}
+                          />
+                          <TextField
+                            error={Boolean(formik.touched.jobTitle && formik.errors.jobTitle)}
+                            fullWidth
+                            helperText={formik.touched.jobTitle && formik.errors.jobTitle}
+                            label="Job title"
+                            name="jobTitle"
+                            onBlur={formik.handleBlur}
+                            onChange={formik.handleChange}
+                            value={formik.values.jobTitle}
+                          />
+                          <TextField
+                            error={Boolean(formik.touched.companyName && formik.errors.companyName)}
+                            fullWidth
+                            helperText={formik.touched.companyName && formik.errors.companyName}
+                            label="Company name"
+                            name="companyName"
+                            onBlur={formik.handleBlur}
+                            onChange={formik.handleChange}
+                            value={formik.values.companyName}
+                          />
+                          <TextField
+                            error={Boolean(formik.touched.companySize && formik.errors.companySize)}
+                            fullWidth
+                            helperText={formik.touched.companySize && formik.errors.companySize}
+                            label="Company size"
+                            name="companySize"
+                            onBlur={formik.handleBlur}
+                            onChange={formik.handleChange}
+                            select
+                            value={formik.values.companySize}
+                          >
+                            {companySizeOptions.map((companySizeOption) => (
+                              <MenuItem
+                                key={companySizeOption}
+                                value={companySizeOption}
+                              >
+                                {companySizeOption}
+                              </MenuItem>
+                            ))}
+                          </TextField>
+                        </Stack>
+                        {formik.errors.submit && (
+                          <FormHelperText
+                            error
+                            sx={{ mt: 3 }}
+                          >
                             {formik.errors.submit}
                           </FormHelperText>
-                        </Grid>
-                      )}
-                      <Grid
-                        item
-                        xs={12}
-                      >
-                        <Button
-                          color="primary"
-                          size="large"
-                          type="submit"
-                          variant="contained"
-                        >
-                          Save settings
-                        </Button>
-                      </Grid>
-                    </Grid>
-                  </div>
-                </form>
-              </Card>
-            </Grid>
-          </Grid>
+                        )}
+                        <Box sx={{ mt: 3 }}>
+                          <Button
+                            color="primary"
+                            size="large"
+                            type="submit"
+                            variant="contained"
+                          >
+                            Save settings
+                          </Button>
+                        </Box>
+                      </Box>
+                    </form>
+                  </Card>
+                </Grid>
+              </Grid>
+            </div>
+          </Stack>
         </Container>
       </Box>
     </>
   );
 };
+
+export default Page;
